requirements.txt
requests
hvac
python-dotenv
strands-agents[ollama]